backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "llama-index-core>=0.12,<0.13",
    "llama-index-embeddings-openai>=0.3,<0.5",
    "llama-index-llms-openrouter>=0.3,<0.5",
    "llama-index-readers-file>=0.4,<0.5",
    "llama-index-retrievers-bm25>=0.5,<0.6",
    "pydantic-settings>=2.0,<3.0",
    "sentence-transformers>=3.0,<4.0",
    "pytest>=9.0.2",
    "python-dotenv==1.1.0",
    "sse-starlette==2.2.1",
    "uvicorn[standard]==0.34.2",
]
